fully using that a stack has no gaps and has same resolution. integer indexing wherever possible. loaders totally changed. changed displacement XY to be in um, as for Z

+++ ev/endrov/modelWindowImset/slice3d/Slice3D.java
@@ -90,7 +90,7 @@ public class Slice3D
 				zplane=0;
 			if(zplane>stack.getDepth())
 				zplane=stack.getDepth();
-			worldZ=zplane*stack.resZ.doubleValue();
+			worldZ=zplane*stack.resZ;
 			EvImage evim=stack.getInt(zplane);
 			EvPixels p=evim.getPixels();
 			w=p.getWidth();
